requirements.txt
python-dotenv
zeep
xmlschema
